Struct ic_web3::api::Eth

pub struct Eth<T> { /* private fields */ }

Eth namespace

Implementations

impl<T: Transport> Eth<T>

pub fn accounts(&self) -> CallFuture<Vec<Address>, T::Out>

Get list of available accounts.

pub fn block_number(&self) -> CallFuture<U64, T::Out>

Get current block number

pub fn call( &self, req: CallRequest, block: Option<BlockId> ) -> CallFuture<Bytes, T::Out>

Call a constant method of contract without changing the state of the blockchain.

pub fn coinbase(&self) -> CallFuture<Address, T::Out>

Get coinbase address

pub fn compile_lll(&self, code: String) -> CallFuture<Bytes, T::Out>

Compile LLL

pub fn compile_solidity(&self, code: String) -> CallFuture<Bytes, T::Out>

Compile Solidity

pub fn compile_serpent(&self, code: String) -> CallFuture<Bytes, T::Out>

Compile Serpent

pub fn estimate_gas( &self, req: CallRequest, block: Option<BlockNumber> ) -> CallFuture<U256, T::Out>

Call a contract without changing the state of the blockchain to estimate gas usage.

pub fn gas_price(&self) -> CallFuture<U256, T::Out>

Get current recommended gas price

pub fn fee_history( &self, block_count: U256, newest_block: BlockNumber, reward_percentiles: Option<Vec<f64>> ) -> CallFuture<FeeHistory, T::Out>

Returns a collection of historical gas information. This can be used for evaluating the max_fee_per_gas and max_priority_fee_per_gas to send the future transactions.

pub fn balance( &self, address: Address, block: Option<BlockNumber> ) -> CallFuture<U256, T::Out>

Get balance of given address

pub fn logs(&self, filter: Filter) -> CallFuture<Vec<Log>, T::Out>

Get all logs matching a given filter object

pub fn block(&self, block: BlockId) -> CallFuture<Option<Block<H256>>, T::Out>

Get block details with transaction hashes.

pub fn block_with_txs( &self, block: BlockId ) -> CallFuture<Option<Block<Transaction>>, T::Out>

Get block details with full transaction objects.

pub fn block_transaction_count( &self, block: BlockId ) -> CallFuture<Option<U256>, T::Out>

Get number of transactions in block

pub fn code( &self, address: Address, block: Option<BlockNumber> ) -> CallFuture<Bytes, T::Out>

Get code under given address

pub fn compilers(&self) -> CallFuture<Vec<String>, T::Out>

Get supported compilers

pub fn chain_id(&self) -> CallFuture<U256, T::Out>

Get chain id

pub fn request_accounts(&self) -> CallFuture<Vec<Address>, T::Out>

Get available user accounts. This method is only available in the browser. With MetaMask, this will cause the popup that prompts the user to allow or deny access to their accounts to your app.

pub fn storage( &self, address: Address, idx: U256, block: Option<BlockNumber> ) -> CallFuture<H256, T::Out>

Get storage entry

pub fn transaction_count( &self, address: Address, block: Option<BlockNumber> ) -> CallFuture<U256, T::Out>

Get nonce

pub fn transaction( &self, id: TransactionId ) -> CallFuture<Option<Transaction>, T::Out>

Get transaction

pub fn transaction_receipt( &self, hash: H256 ) -> CallFuture<Option<TransactionReceipt>, T::Out>

Get transaction receipt

pub fn uncle_header( &self, block: BlockId, index: Index ) -> CallFuture<Option<BlockHeader>, T::Out>

Get uncle header by block ID and uncle index.

This method is meant for TurboGeth compatiblity, which is missing transaction hashes in the response.

pub fn uncle( &self, block: BlockId, index: Index ) -> CallFuture<Option<Block<H256>>, T::Out>

Get uncle by block ID and uncle index – transactions only has hashes.

pub fn uncle_count(&self, block: BlockId) -> CallFuture<Option<U256>, T::Out>

Get uncle count in block

pub fn work(&self) -> CallFuture<Work, T::Out>

Get work package

pub fn hashrate(&self) -> CallFuture<U256, T::Out>

Get hash rate

pub fn mining(&self) -> CallFuture<bool, T::Out>

Get mining status

pub fn new_block_filter(&self) -> CallFuture<U256, T::Out>

Start new block filter

pub fn new_pending_transaction_filter(&self) -> CallFuture<U256, T::Out>

Start new pending transaction filter

pub fn protocol_version(&self) -> CallFuture<String, T::Out>

Start new pending transaction filter

pub fn send_raw_transaction(&self, rlp: Bytes) -> CallFuture<H256, T::Out>

Sends a rlp-encoded signed transaction

pub fn send_transaction( &self, tx: TransactionRequest ) -> CallFuture<H256, T::Out>

Sends a transaction transaction

pub fn sign(&self, address: Address, data: Bytes) -> CallFuture<H520, T::Out>

Signs a hash of given data

pub fn submit_hashrate(&self, rate: U256, id: H256) -> CallFuture<bool, T::Out>

Submit hashrate of external miner

pub fn submit_work( &self, nonce: H64, pow_hash: H256, mix_hash: H256 ) -> CallFuture<bool, T::Out>

Submit work of external miner

pub fn syncing(&self) -> CallFuture<SyncState, T::Out>

Get syncing status

pub fn proof( &self, address: Address, keys: Vec<U256>, block: Option<BlockNumber> ) -> CallFuture<Option<Proof>, T::Out>

Returns the account- and storage-values of the specified account including the Merkle-proof.