Struct ic_web3::api::Web3

pub struct Web3<T: Transport> { /* private fields */ }

Web3 wrapper for all namespaces

Implementations

impl<T: Transport> Web3<T>

pub fn new(transport: T) -> Self

Create new Web3 with given transport

pub fn transport(&self) -> &T

Borrows a transport.

pub fn set_max_response_bytes(&mut self, bytes: u64)

set the max response bytes

pub fn api<A: Namespace<T>>(&self) -> A

Access methods from custom namespace

pub fn accounts(&self) -> Accounts<T>

Access methods from accounts namespace

pub fn eth(&self) -> Eth<T>

Access methods from eth namespace

pub fn net(&self) -> Net<T>

Access methods from net namespace

pub fn web3(&self) -> Web3<T>

Access methods from web3 namespace

pub fn eth_filter(&self) -> EthFilter<T>

Access filter methods from eth namespace

pub fn parity(&self) -> Parity<T>

Access methods from parity namespace

pub fn parity_accounts(&self) -> ParityAccounts<T>

Access methods from parity_accounts namespace

pub fn parity_set(&self) -> ParitySet<T>

Access methods from parity_set namespace

pub fn personal(&self) -> Personal<T>

Access methods from personal namespace

pub fn trace(&self) -> Traces<T>

Access methods from trace namespace

pub fn txpool(&self) -> Txpool<T>

Access methods from txpool namespace

pub async fn wait_for_confirmations<F, V>( &self, poll_interval: Duration, confirmations: usize, check: V ) -> Result<()>where F: Future<Output = Result<Option<U64>>>, V: ConfirmationCheck<Check = F>,

Should be used to wait for confirmations

pub async fn send_transaction_with_confirmation( &self, tx: TransactionRequest, poll_interval: Duration, confirmations: usize ) -> Result<TransactionReceipt>

Sends transaction and returns future resolved after transaction is confirmed

pub async fn send_raw_transaction_with_confirmation( &self, tx: Bytes, poll_interval: Duration, confirmations: usize ) -> Result<TransactionReceipt>

Sends raw transaction and returns future resolved after transaction is confirmed

pub async fn json_rpc_call(&self, body: &str) -> Result<String>

Call json rpc directly

impl<T: DuplexTransport> Web3<T>

pub fn eth_subscribe(&self) -> EthSubscribe<T>

Access subscribe methods from eth namespace