Struct ic_web3::contract::Contract

pub struct Contract<T: Transport> { /* private fields */ }

Ethereum Contract Interface

Implementations

impl<T: Transport> Contract<T>

pub async fn sign( &self, func: &str, params: impl Tokenize, options: Options, from: String, key_info: KeyInfo, chain_id: u64 ) -> Result<SignedTransaction>

pub async fn signed_call( &self, func: &str, params: impl Tokenize, options: Options, from: String, key_info: KeyInfo, chain_id: u64 ) -> Result<H256>

Submit contract call transaction to the transaction pool.

Note this function DOES NOT wait for any confirmations, so there is no guarantees that the call is actually executed. If you’d rather wait for block inclusion, please use [signed_call_with_confirmations] instead.

pub async fn signed_call_with_confirmations( &self, func: &str, params: impl Tokenize, options: Options, from: String, confirmations: usize, key_info: KeyInfo, chain_id: u64 ) -> Result<TransactionReceipt>

impl<T: Transport> Contract<T>

pub fn deploy(eth: Eth<T>, json: &[u8]) -> Result<Builder<T>>

Creates deployment builder for a contract given it’s ABI in JSON.

pub fn deploy_from_truffle<S>( eth: Eth<T>, json: &[u8], linker: HashMap<S, Address> ) -> Result<Builder<T>>where S: AsRef<str> + Eq + Hash,

test

impl<T: Transport> Contract<T>

pub fn new(eth: Eth<T>, address: Address, abi: Contract) -> Self

Creates new Contract Interface given blockchain address and ABI

pub fn from_json(eth: Eth<T>, address: Address, json: &[u8]) -> Result<Self>

Creates new Contract Interface given blockchain address and JSON containing ABI

pub fn abi(&self) -> &Contract

Get the underlying contract ABI.

pub fn address(&self) -> Address

Returns contract address

pub async fn call<P>( &self, func: &str, params: P, from: Address, options: Options ) -> Result<H256>where P: Tokenize,

Execute a contract function

pub async fn call_with_confirmations( &self, func: &str, params: impl Tokenize, from: Address, options: Options, confirmations: usize ) -> Result<TransactionReceipt>

Execute a contract function and wait for confirmations

pub async fn estimate_gas<P>( &self, func: &str, params: P, from: Address, options: Options ) -> Result<U256>where P: Tokenize,

Estimate gas required for this function call.

pub fn query<R, A, B, P>( &self, func: &str, params: P, from: A, options: Options, block: B ) -> impl Future<Output = Result<R>> + '_where R: Detokenize, A: Into<Option<Address>>, B: Into<Option<BlockId>>, P: Tokenize,

Call constant function

pub async fn events<A, B, C, R>( &self, event: &str, topic0: A, topic1: B, topic2: C ) -> Result<Vec<R>>where A: Tokenize, B: Tokenize, C: Tokenize, R: Detokenize,

Find events matching the topics.