Trait ic_web3::Transport

pub trait Transport: Debug + Clone {
    type Out: Future<Output = Result<Value>>;

    // Required methods
    fn prepare(&self, method: &str, params: Vec<Value>) -> (RequestId, Call);
    fn send(&self, id: RequestId, request: Call) -> Self::Out;

    // Provided methods
    fn execute(&self, method: &str, params: Vec<Value>) -> Self::Out { ... }
    fn set_max_response_bytes(&mut self, bytes: u64) { ... }
}

Transport implementation

Required Associated Types

type Out: Future<Output = Result<Value>>

The type of future this transport returns when a call is made.

Required Methods

fn prepare(&self, method: &str, params: Vec<Value>) -> (RequestId, Call)

Prepare serializable RPC call for given method with parameters.

fn send(&self, id: RequestId, request: Call) -> Self::Out

Execute prepared RPC call.

Provided Methods

fn execute(&self, method: &str, params: Vec<Value>) -> Self::Out

Execute remote method with given parameters.

fn set_max_response_bytes(&mut self, bytes: u64)

set the max response bytes, do nothing by default